+++ crates/pytauri-core/src/ext_mod_impl/lib/mod.rs
@@ -8,4 +8,5 @@ pub(crate) mod listener;
 pub(crate) mod manager;
 pub(crate) mod rect;
 pub(crate) mod run_event;
+pub(crate) mod theme;
 pub(crate) mod url;