+++ crates/pytauri-core/src/ext_mod_impl/lib/run_event.rs
@@ -1,13 +1,18 @@
+use std::path::PathBuf;
+
 use pyo3::{
-    exceptions::PyNotImplementedError,
     prelude::*,
-    types::{PyInt, PyString},
-    IntoPyObject,
+    types::{PyBool, PyFloat, PyInt, PyList, PyString},
+    BoundObject as _, IntoPyObject,
+};
+use pyo3_utils::py_wrapper::{PyWrapper, PyWrapperT0};
+
+use crate::ext_mod::{
+    menu::MenuEvent, tray::TrayIconEvent, PhysicalPositionF64, PhysicalPositionI32,
+    PhysicalSizeU32, Theme,
 };
 
-use crate::ext_mod::{menu::MenuEvent, tray::TrayIconEvent};
-
-/// see also: [tauri::RunEvent]
+/// See also: [tauri::RunEvent]
 #[pyclass(frozen)]
 #[non_exhaustive]
 pub enum RunEvent {
@@ -17,22 +22,19 @@ pub enum RunEvent {
     #[non_exhaustive]
     ExitRequested {
         code: Option<Py<PyInt>>,
-        // TODO, XXX, FIXME: `ExitRequestApi` is a private type in `tauri`,
-        // we need create a issue to `tauri`, or we cant implement this.
-        // See: <https://github.com/tauri-apps/tauri/pull/12701>
-        // api: ExitRequestApi,
+        api: Py<ExitRequestApi>,
     },
+    // TODO, PERF: maybe we should remove `WindowEvent` and `WebviewEvent` fields,
+    // use `on_window_event` and `on_webview_event` instead.
     #[non_exhaustive]
     WindowEvent {
         label: Py<PyString>,
-        // TODO:
-        // event: WindowEvent,
+        event: Py<WindowEvent>,
     },
     #[non_exhaustive]
     WebviewEvent {
         label: Py<PyString>,
-        // TODO:
-        // event: WebviewEvent,
+        event: Py<WebviewEvent>,
     },
     Ready(),
     Resumed(),
@@ -41,31 +43,34 @@ pub enum RunEvent {
     // use `on_menu_event` and `on_tray_icon_event` instead.
     MenuEvent(Py<MenuEvent>),
     TrayIconEvent(Py<TrayIconEvent>),
+    _NonExhaustive(),
 }
 
 impl RunEvent {
-    pub(crate) fn new(py: Python<'_>, value: tauri::RunEvent) -> PyResult<Self> {
+    pub(crate) fn from_tauri(py: Python<'_>, value: tauri::RunEvent) -> PyResult<Self> {
         let ret = match value {
             tauri::RunEvent::Exit => Self::Exit(),
-            tauri::RunEvent::ExitRequested {
-                code, /* TODO */ ..
-            } => {
+            tauri::RunEvent::ExitRequested { code, api, .. } => {
                 let code = code.map(|code| {
                     let Ok(code) = code.into_pyobject(py);
                     code.unbind()
                 });
-                Self::ExitRequested { code }
+                let api = ExitRequestApi::new(api).into_pyobject(py)?.unbind();
+                Self::ExitRequested { code, api }
             }
-            tauri::RunEvent::WindowEvent {
-                label, /* TODO */ ..
-            } => Self::WindowEvent {
-                // if `label` is immutable, we can intern it to save memory.
+            tauri::RunEvent::WindowEvent { label, event, .. } => Self::WindowEvent {
+                // PERF: if `label` is immutable, we can intern it to save memory.
                 label: PyString::intern(py, &label).unbind(),
+                event: WindowEvent::from_tauri(py, &event)?
+                    .into_pyobject(py)?
+                    .unbind(),
             },
-            tauri::RunEvent::WebviewEvent {
-                label, /* TODO */ ..
-            } => Self::WebviewEvent {
+            tauri::RunEvent::WebviewEvent { label, event, .. } => Self::WebviewEvent {
+                // PERF: if `label` is immutable, we can intern it to save memory.
                 label: PyString::intern(py, &label).unbind(),
+                event: WebviewEvent::from_tauri(py, &event)?
+                    .into_pyobject(py)?
+                    .unbind(),
             },
             tauri::RunEvent::Ready => Self::Ready(),
             tauri::RunEvent::Resumed => Self::Resumed(),
@@ -74,15 +79,208 @@ impl RunEvent {
                 Self::MenuEvent(MenuEvent::intern(py, &event.id.0).unbind())
             }
             tauri::RunEvent::TrayIconEvent(event) => Self::TrayIconEvent(
-                TrayIconEvent::from_tauri(py, event)?
+                TrayIconEvent::from_tauri(py, &event)?
                     .into_pyobject(py)?
                     .unbind(),
             ),
-            event => {
-                return Err(PyNotImplementedError::new_err(format!(
-                    "Please make a issue for unimplemented RunEvent: {event:?}",
-                )))
+            _ => Self::_NonExhaustive(),
+        };
+        Ok(ret)
+    }
+}
+
+/// See also: [tauri::CloseRequestApi]
+#[pyclass(frozen)]
+#[non_exhaustive]
+pub struct CloseRequestApi(pub PyWrapper<PyWrapperT0<tauri::CloseRequestApi>>);
+
+impl CloseRequestApi {
+    fn new(value: tauri::CloseRequestApi) -> Self {
+        Self(PyWrapper::new0(value))
+    }
+}
+
+#[pymethods]
+impl CloseRequestApi {
+    // PERF: [Sender::send] is quick enough and never blocks,
+    // so we don't need to release the GIL.
+    fn prevent_close(&self) {
+        self.0.inner_ref().prevent_close();
+    }
+}
+
+/// See also: [tauri::ExitRequestApi]
+#[pyclass(frozen)]
+#[non_exhaustive]
+pub struct ExitRequestApi(pub PyWrapper<PyWrapperT0<tauri::ExitRequestApi>>);
+
+impl ExitRequestApi {
+    fn new(value: tauri::ExitRequestApi) -> Self {
+        Self(PyWrapper::new0(value))
+    }
+}
+
+#[pymethods]
+impl ExitRequestApi {
+    // PERF: [Sender::send] is quick enough and never blocks,
+    // so we don't need to release the GIL.
+    fn prevent_exit(&self) {
+        self.0.inner_ref().prevent_exit();
+    }
+}
+
+/// See also: [tauri::DragDropEvent::Enter::paths]
+///
+/// `list[pathlib.Path]`
+#[derive(FromPyObject, IntoPyObject, IntoPyObjectRef)]
+#[pyo3(transparent)]
+struct VecPathBuf(Py<PyList>);
+
+impl VecPathBuf {
+    #[inline]
+    fn from_tauri(py: Python<'_>, paths: &Vec<PathBuf>) -> PyResult<Self> {
+        let lst = PyList::new(py, paths)?;
+        Ok(Self(lst.unbind()))
+    }
+}
+
+/// See also: [tauri::DragDropEvent]
+#[pyclass(frozen)]
+#[non_exhaustive]
+pub enum DragDropEvent {
+    // use `Py<T>` to avoid creating new obj every time visiting the field,
+    // see: <https://pyo3.rs/v0.23.4/faq.html#pyo3get-clones-my-field>
+    Enter {
+        #[expect(private_interfaces)]
+        paths: VecPathBuf,
+        #[expect(private_interfaces)]
+        position: PhysicalPositionF64,
+    },
+    Over {
+        #[expect(private_interfaces)]
+        position: PhysicalPositionF64,
+    },
+    Drop {
+        #[expect(private_interfaces)]
+        paths: VecPathBuf,
+        #[expect(private_interfaces)]
+        position: PhysicalPositionF64,
+    },
+    Leave(),
+    _NonExhaustive(),
+}
+
+impl DragDropEvent {
+    fn from_tauri(py: Python<'_>, value: &tauri::DragDropEvent) -> PyResult<Self> {
+        let ret = match value {
+            tauri::DragDropEvent::Enter { paths, position } => Self::Enter {
+                paths: VecPathBuf::from_tauri(py, paths)?,
+                position: PhysicalPositionF64::from_tauri(py, *position)?,
+            },
+            tauri::DragDropEvent::Over { position } => Self::Over {
+                position: PhysicalPositionF64::from_tauri(py, *position)?,
+            },
+            tauri::DragDropEvent::Drop { paths, position } => Self::Drop {
+                paths: VecPathBuf::from_tauri(py, paths)?,
+                position: PhysicalPositionF64::from_tauri(py, *position)?,
+            },
+            tauri::DragDropEvent::Leave => Self::Leave(),
+            _ => Self::_NonExhaustive(),
+        };
+        Ok(ret)
+    }
+}
+
+/// See also: [tauri::WebviewEvent]
+#[pyclass(frozen)]
+#[non_exhaustive]
+pub enum WebviewEvent {
+    // use `Py<T>` to avoid creating new obj every time visiting the field,
+    // see: <https://pyo3.rs/v0.23.4/faq.html#pyo3get-clones-my-field>
+    DragDrop(Py<DragDropEvent>),
+    _NonExhaustive(),
+}
+
+impl WebviewEvent {
+    pub(crate) fn from_tauri(py: Python<'_>, value: &tauri::WebviewEvent) -> PyResult<Self> {
+        let ret = match value {
+            tauri::WebviewEvent::DragDrop(event) => Self::DragDrop(
+                DragDropEvent::from_tauri(py, event)?
+                    .into_pyobject(py)?
+                    .unbind(),
+            ),
+            _ => Self::_NonExhaustive(),
+        };
+        Ok(ret)
+    }
+}
+
+/// See also: [tauri::WindowEvent]
+#[pyclass(frozen)]
+#[non_exhaustive]
+pub enum WindowEvent {
+    // use `Py<T>` to avoid creating new obj every time visiting the field,
+    // see: <https://pyo3.rs/v0.23.4/faq.html#pyo3get-clones-my-field>
+    #[expect(private_interfaces)]
+    Resized(PhysicalSizeU32),
+    #[expect(private_interfaces)]
+    Moved(PhysicalPositionI32),
+    #[non_exhaustive]
+    CloseRequested {
+        api: Py<CloseRequestApi>,
+    },
+    Destroyed(),
+    Focused(Py<PyBool>),
+    #[non_exhaustive]
+    ScaleFactorChanged {
+        scale_factor: Py<PyFloat>,
+        #[expect(private_interfaces)]
+        new_inner_size: PhysicalSizeU32,
+    },
+    DragDrop(Py<DragDropEvent>),
+    ThemeChanged(Py<Theme>),
+    _NonExhaustive(),
+}
+
+impl WindowEvent {
+    // NOTE: Because the parameter of [tauri::webview::WebviewWindow::on_window_event] is `&WindowEvent`,
+    // and we do not want to clone [tauri::WindowEvent::DragDrop] (since it contains [Vec<PathBuf>]),
+    // we use `&tauri::WindowEvent` as the parameter here.
+    pub(crate) fn from_tauri(py: Python<'_>, value: &tauri::WindowEvent) -> PyResult<Self> {
+        let ret = match value {
+            tauri::WindowEvent::Resized(size) => {
+                Self::Resized(PhysicalSizeU32::from_tauri(py, *size)?)
             }
+            tauri::WindowEvent::Moved(pos) => {
+                Self::Moved(PhysicalPositionI32::from_tauri(py, *pos)?)
+            }
+            tauri::WindowEvent::CloseRequested { api, .. } => {
+                let api = CloseRequestApi::new(api.clone())
+                    .into_pyobject(py)?
+                    .unbind();
+                Self::CloseRequested { api }
+            }
+            tauri::WindowEvent::Destroyed => Self::Destroyed(),
+            tauri::WindowEvent::Focused(focused) => {
+                Self::Focused(PyBool::new(py, *focused).unbind())
+            }
+            tauri::WindowEvent::ScaleFactorChanged {
+                scale_factor,
+                new_inner_size,
+                ..
+            } => Self::ScaleFactorChanged {
+                scale_factor: PyFloat::new(py, *scale_factor).unbind(),
+                new_inner_size: PhysicalSizeU32::from_tauri(py, *new_inner_size)?,
+            },
+            tauri::WindowEvent::DragDrop(event) => Self::DragDrop(
+                DragDropEvent::from_tauri(py, event)?
+                    .into_pyobject(py)?
+                    .unbind(),
+            ),
+            tauri::WindowEvent::ThemeChanged(theme) => {
+                Self::ThemeChanged(Theme::from(*theme).into_pyobject(py)?.unbind())
+            }
+            _ => Self::_NonExhaustive(),
         };
         Ok(ret)
     }